+++ app/src/main/java/ca/stefanm/sayhi/MainActivity.java
@@ -1,0 +1,49 @@
+package ca.stefanm.sayhi;
+
+
+
+
+import android.app.Activity;
+import android.app.FragmentTransaction;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+
+import ca.stefanm.sayhi.ui.SettingsFragment;
+
+public class MainActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.action_settings) {
+            FragmentTransaction transaction = getFragmentManager().beginTransaction();
+            //transaction.show(new SettingsFragment());
+            transaction.replace(R.id.fragment, new SettingsFragment());
+            transaction.addToBackStack(null);
+            transaction.commit();
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+}
