+++ app/src/main/java/ca/stefanm/sayhi/ui/NearbyListFragment.java
@@ -1,0 +1,9 @@
+package ca.stefanm.sayhi.ui;
+
+import android.app.Fragment;
+
+/**
+ * Created by stefan on 8/8/15.
+ */
+public class NearbyListFragment extends Fragment {
+}
