+++ app/src/main/res/layout/activity_main.xml
@@ -1,0 +1,5 @@
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:id="@+id/fragment"
+    android:name="ca.stefanm.sayhi.MainActivityFragment" tools:layout="@layout/fragment_main"
+    android:layout_width="match_parent" android:layout_height="match_parent"
+    tools:context=".MainActivity" />
