+++ app/src/main/res/layout/fragment_main.xml
@@ -4,14 +4,19 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".HomeActivity$PlaceholderFragment">
+    tools:context="ca.stefanm.sayhi.MainActivityFragment">
 
-    <ListView
+    <TextView android:text="@string/hello_world" android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/textView2" />
+
+    <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/listView"
-        android:layout_alignParentTop="true"
+        android:text="New Button"
+        android:id="@+id/button"
+        android:layout_below="@+id/textView2"
         android:layout_alignParentStart="true"
-        tools:listitem="@layout/nearbyperson_listitem" />
+        android:layout_marginStart="32dp" />
 
 </RelativeLayout>
