+++ app/src/main/res/values/strings.xml
@@ -14,4 +14,6 @@
     <string name="title_hamburgermenu_home">Home</string>
     <string name="title_hamburgermenu_freq_contacted">Frequently Contacted</string>
     <string name="title_hamburgermenu_settings">Settings</string>
+    <string name="title_activity_main">MainActivity</string>
+    <string name="hello_world">Hello world!</string>
 </resources>
